Add PDF Adapter
