+++ kkh-handbook/Models/File.swift
@@ -8,7 +8,6 @@
 
 import Foundation
 import UIKit
-import PDFReader
 
 struct File {
 	let name: String
@@ -18,13 +17,6 @@ struct File {
 	
 	// Display PDFViewer
 	func buildController() -> UIViewController  {
-		return PDFViewController.createNew(with: PDFAdapter.displayFile(name))
-	}
-}
-
-class PDFAdapter {
-	static func displayFile (_ name: String) -> PDFDocument {
-		let documentFileURL = Bundle.main.url(forResource: name, withExtension: "pdf")!
-		return PDFDocument(url: documentFileURL)!
+		return PDFAdapter.buildView(self.name)
 	}
 }
